separate navbar for admin and user
