+++ frontend/jsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "baseUrl": "src",
     "paths": {
-      "*": ["src/*"],
+      "*": ["src/routes/*"],
       "assets": ["src/assets"]
     }
   }
