+++ frontend/src/layouts/Admin.js
@@ -10,7 +10,7 @@ import DemoNavbar from "components/Navbars/DemoNavbar.js";
 import Footer from "components/Footer/Footer.js";
 import Sidebar from "components/Sidebar/Sidebar.js";
 
-import routes from "routes.js";
+import routes from "routes/admin.js";
 
 var ps;
